BidirectionalMap.Serialization.Json were retargeted to .NET Standard 1.6.
BidirectionalMap project reference was replaced with NuGet package reference.

System.Text.Json was replaced with Newtonsoft.Json because System.Text.Json is for .NET Standard 2.0+.

BidirectionalMap.Serialization.Json with BidirectionalMap.Serialization.Json.Test were renamed to BidirectionalMap.Json.Newtonsoft and BidirectionalMap.Json.Newtonsoft.Test.

All this has been done because only the old BiMap API requires a special JSON converter.
