+++ src/BidirectionalMap.Json.Newtonsoft/BidirectionalMap.Json.Newtonsoft.csproj
@@ -1,0 +1,17 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard1.6</TargetFramework>
+    <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
+    <Authors>Spencer Farley, Ihor Kuzmychov</Authors>
+    <Description>Provides JSON Newtonsoft-converter for BidirectionalMap v1.0.0.</Description>
+    <PackageProjectUrl>https://github.com/farlee2121/BidirectionalMap</PackageProjectUrl>
+    <PackageTags>dotnet-standard;lookup;map;bidirectional;dictionary;bimap;json;newtonsoft</PackageTags>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="BidirectionalMap" Version="1.0.0" />
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+  </ItemGroup>
+
+</Project>
