--- src/BidirectionalMap.Serialization.Json/BidirectionalMap.Serialization.Json.csproj
@@ -1,16 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.1</TargetFramework>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="System.Text.Json" Version="6.0.5" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\BidirectionalMap\BidirectionalMap.csproj" />
-  </ItemGroup>
-
-</Project>
