+++ tests/BidirectionalMap.Json.Newtonsoft.Tests/BidirectionalMap.Json.Newtonsoft.Tests.csproj
@@ -13,10 +13,11 @@
     <PackageReference Include="MSTest.TestAdapter" Version="2.2.8" />
     <PackageReference Include="MSTest.TestFramework" Version="2.2.8" />
     <PackageReference Include="coverlet.collector" Version="3.1.2" />
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\BidirectionalMap.Serialization.Json\BidirectionalMap.Serialization.Json.csproj" />
+    <ProjectReference Include="..\..\src\BidirectionalMap.Json.Newtonsoft\BidirectionalMap.Json.Newtonsoft.csproj" />
   </ItemGroup>
 
 </Project>
